Use v2 module, keeping v1 for server commands at least
