+++ deprecated/fabric-commands-v0/build.gradle
@@ -3,5 +3,5 @@ version = getSubprojectVersion(project)
 
 moduleDependencies(project, [
 		'fabric-api-base',
-		'fabric-command-api-v1'
+		'fabric-command-api-v2'
 ])
