+++ fabric-command-api-v1/build.gradle
@@ -1,10 +1,7 @@
 archivesBaseName = "fabric-command-api-v1"
 version = getSubprojectVersion(project)
 
-dependencies {
-	testmodImplementation project(path: ':fabric-lifecycle-events-v1', configuration: 'namedElements')
-}
-
 moduleDependencies(project, [
-		'fabric-api-base'
+		'fabric-api-base',
+		'fabric-command-api-v2'
 ])
