--- fabric-command-api-v1/src/testmod/resources/data/minecraft/tags/functions/load.json
@@ -1,6 +1,0 @@
-{
-  "replace": false,
-  "values":[
-    "fabric_command_api_v1_testmod:load"
-  ]
-}
