+++ fabric-command-api-v2/build.gradle
@@ -1,0 +1,10 @@
+archivesBaseName = "fabric-command-api-v2"
+version = getSubprojectVersion(project)
+
+dependencies {
+	testmodImplementation project(path: ':fabric-lifecycle-events-v1', configuration: 'namedElements')
+}
+
+moduleDependencies(project, [
+		'fabric-api-base'
+])
