+++ fabric-command-api-v2/src/main/java/net/fabricmc/fabric/mixin/command/client/ClientPlayerEntityMixin.java
