+++ fabric-command-api-v2/src/main/resources/fabric-command-api-v2.mixins.json
