+++ fabric-command-api-v2/src/testmod/resources/data/fabric_command_api_v2_testmod/functions/load.mcfunction
