+++ fabric-command-api-v2/src/testmod/resources/data/minecraft/tags/functions/tick.json
@@ -1,0 +1,6 @@
+{
+  "replace": false,
+  "values":[
+    "fabric_command_api_v2_testmod:tick"
+  ]
+}
