+++ fabric-entity-events-v1/src/testmod/resources/fabric.mod.json
@@ -7,7 +7,7 @@
   "license": "Apache-2.0",
   "depends": {
     "fabric-entity-events-v1": "*",
-    "fabric-command-api-v1": "*",
+    "fabric-command-api-v2": "*",
     "fabric-registry-sync-v0": "*"
   },
   "entrypoints": {
