+++ gradle.properties
@@ -13,6 +13,7 @@ fabric-api-lookup-api-v1-version=1.6.2
 fabric-biome-api-v1-version=9.0.6
 fabric-blockrenderlayer-v1-version=1.1.14
 fabric-command-api-v1-version=1.1.17
+fabric-command-api-v2-version=1.0.0
 fabric-commands-v0-version=0.2.16
 fabric-containers-v0-version=0.1.23
 fabric-content-registries-v0-version=3.0.10
