+++ settings.gradle
@@ -17,6 +17,7 @@ include 'fabric-api-lookup-api-v1'
 include 'fabric-biome-api-v1'
 include 'fabric-blockrenderlayer-v1'
 include 'fabric-command-api-v1'
+include 'fabric-command-api-v2'
 include 'fabric-content-registries-v0'
 include 'fabric-crash-report-info-v1'
 include 'fabric-data-generation-api-v1'
